requirements.txt
tweet-preprocessor==0.6.0
transformers==3.1.0
pyarrow==1.0.1
torch==1.6.0
nlp==0.4.0
